cleaned up codes from comments and debugger
