+++ Makefile
@@ -24,6 +24,7 @@ SRC =	./src/builtin.c \
 		./src/main.c \
 		./src/parse.c \
 		./src/parse_util.c \
+		./src/parse_util2.c \
 		./src/pipes.c \
 		./src/prompt_check.c \
 		./src/safe_malloc.c \
@@ -31,6 +32,7 @@ SRC =	./src/builtin.c \
 		./src/unquotes.c \
 		./src/utils.c \
 		./src/utils_helpers.c \
+		./src/utils_helpers2.c \
 		./src/safe_func.c \
 
 #BSRC =
