+++ src/here_doc.c
@@ -3,19 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   here_doc.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jlu <jlu@student.42.fr>                    +#+  +:+       +#+        */
+/*   By: jlu <jlu@student.hive.fi>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/19 23:17:23 by jlu               #+#    #+#             */
-/*   Updated: 2024/07/15 19:09:12 by jlu              ###   ########.fr       */
+/*   Updated: 2024/07/16 14:56:22 by jlu              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
 int g_sigint = 0;
-
-
-/* check for here_doc if it returns -1, no here_doc if its >= 0 then it means here doc found*/
 
 int    check_heredoc(t_token *t)
 {
@@ -35,78 +32,6 @@ int    check_heredoc(t_token *t)
     }
     return (0);
 }
-// // when i exit with ctrl-d, the bad address caused leaks
-// int    ft_hd(t_data *data, int i, int j)
-// {
-//     int hd;
-//     // char *str;
-//     char buf[512];
-//     char tmp_buf[1];
-//     char *limiter;
-//     char *hdfile;
-//     ssize_t bytes;
-//     ssize_t t_bytes;
-
-//     // printf("went to hd\n");
-//     hdfile = ft_itoa(i);
-//     if (!hdfile)
-//         error(data, XMALLOC, EXIT_FAILURE);
-//     limiter = ft_strjoin(data->token[i].redir[j][1], "\n");
-//     hd = open(hdfile, O_CREAT | O_TRUNC | O_WRONLY, 0644);
-// 	if (hd < 0)
-// 		error(data, XHD, EXIT_FAILURE);
-//     // str = ft_safe_malloc(sizeof(char));
-//     set_signal_handler(SIGINT, heredoc_handler);
-//     t_bytes = 0;
-//     while (1)
-//     {
-//         if (t_bytes == 0)
-//             write (1, "> ", 2);
-//         bytes = read(STDIN_FILENO, tmp_buf, 1);
-//         if (bytes < 0)
-//             break ;
-//         if (bytes == 0)
-//         {
-//             if (t_bytes == 0)
-//             {
-//                 printf("\n");
-//                 break ;
-//             }
-//             else
-//                 continue ; 
-//         }
-//         buf[t_bytes] = tmp_buf[0];
-//         t_bytes ++;
-//         buf[t_bytes] = '\0';
-//         if (tmp_buf[0] == '\n')
-//         {
-//             if (!ft_strcmp(buf, limiter))
-//                 break ;
-//             else
-//             {
-//                 // str = check_expand(buf, data);
-//                 ft_putstr_fd(buf, hd);
-//                 t_bytes = 0;
-//             }
-//         }
-//     }
-//     // printf("g_sigint: %d\n", g_sigint);
-//     if (g_sigint == 1)
-//     {
-//         unlink(hdfile);
-//         free(hdfile);
-//         free(limiter);
-//         // free(str);
-//         close (hd);
-//         g_sigint = 0;
-//         return (0);
-//     }
-//     // free(str);
-//     free(hdfile);
-//     free(limiter);
-//     data->token[i].hd = hd;
-//     return (1);
-// }
 
 static int hd_init(t_data *data, int i)
 {
@@ -121,15 +46,17 @@ static int hd_init(t_data *data, int i)
     return (fd);
 }
 
-static void hd_done(t_data *data, int hd, int i)
+static int hd_done(t_data *data, int hd, int i)
 {
     if (g_sigint == 1)
     {
         unlink(data->token[i].hdfile);
         close(hd);
         g_sigint = 0;
+        return (0);
     }
-    // return (0);
+    data->token[i].hd = hd;
+    return (1);
 }
 
 int    ft_hd(t_data *data, int i, int j)
@@ -149,75 +76,17 @@ int    ft_hd(t_data *data, int i, int j)
         if (line && *line != '\0')
         {
             if (!ft_strcmp(line, data->token[i].redir[j][1]))
-            {
-                free(line);
                 break;
-            }
             line = check_expand(line, data);
             ft_putendl_fd(line, hd);
         }
 		free(line);
 	}
+    free(line);
     safe_dup2(stdin_backup, STDIN_FILENO);
-    if (g_sigint == 1)
-    {
-        hd_done(data, hd, i);
-        return (0);
-    }
-    data->token[i].hd = hd;
-    return (1);
+    // data->token[i].hd = hd; //put this in hd_done
+    return (hd_done(data, hd, i));
 }
-
-// int    ft_hd(t_data *data, int i, int j)
-// {
-//     int hd;
-//     char *line;
-//     char *limiter;
-//     char *hdfile;
-
-//     hdfile = ft_itoa(i);
-//     if (!hdfile)
-//         error(data, XMALLOC, EXIT_FAILURE);
-//     data->token[i].hdfile = ft_itoa(i);
-//     if (!data->token[i].hdfile)
-//         error(data, XMALLOC, EXIT_FAILURE);
-//     limiter = data->token[i].redir[j][1];
-//     hd = open(data->token[i].hdfile, O_CREAT | O_TRUNC | O_WRONLY, 0644);
-// 	if (hd < 0)
-// 		error(data, XHD, EXIT_FAILURE);
-//     set_signal_handler(SIGINT, heredoc_handler);
-//     while (g_sigint == 0)
-// 	{
-// 		line = readline("> ");
-//         if (!line)
-//             break ;
-//         if (line)
-//         {
-// 	        if (ft_strcmp(line, limiter))
-//             {
-//                 line = check_expand(line, data);
-//                 ft_putendl_fd(line, hd);
-//             }
-//             else if (!ft_strcmp(line, limiter))
-//             {
-//                 free(line);
-//                 break;
-//             }
-//         }
-// 		free(line);
-// 	}
-//     if (g_sigint == 1)
-//     {
-//         unlink(hdfile);
-//         free(hdfile);
-//         free(limiter);
-//         close (hd);
-//         g_sigint = 0;
-//         return (0);
-//     }
-//     data->token[i].hd = hd;
-//     return (1);
-// }
 
 int    here_doc(t_data *data)
 {
@@ -226,7 +95,6 @@ int    here_doc(t_data *data)
     t_token *t;
 
     t = data->token;
-
     i = 0;
     while (i < data->cmd_count)
     {
