+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jlu <jlu@student.42.fr>                    +#+  +:+       +#+        */
+/*   By: jlu <jlu@student.hive.fi>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/29 15:10:28 by jlu               #+#    #+#             */
-/*   Updated: 2024/07/15 17:26:05 by jlu              ###   ########.fr       */
+/*   Updated: 2024/07/16 13:54:57 by jlu              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,10 +85,9 @@ void	ft_minishell(t_data *data)
 				status = parse_start(data, line);
 			if (status)
 			 	execute(data);
-
+		}
     }
-		free(line);
-	}
+	free(line);
 	toggle_input(SIG_CHILD);
 }
