+++ src/utils.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jlu <jlu@student.hive.fi>                  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/07/16 14:30:18 by jlu               #+#    #+#             */
+/*   Updated: 2024/07/16 14:30:20 by jlu              ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 
 #include "minishell.h"
 
@@ -71,17 +83,3 @@ int	ft_envsize(t_env *lst)
 	}
 	return (count);
 }
-
-// int	ft_strcmp(char *s1, char *s2)
-// {
-// 	int	i;
-// 	int	result;
-
-// 	i = 0;
-// 	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0')
-// 	{
-// 		i++;
-// 	}
-// 	result = (s1[i] - s2[i]);
-// 	return (result);
-// }
